% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apriorimining.R
\name{apriorimining}
\alias{apriorimining}
\title{Determine association rules}
\usage{
apriorimining(input, m_sup, m_conf)
}
\arguments{
\item{input}{Binary matrix containing transaction data, with rows
representing transactions and columns representing items. Can be
either logical or numeric, every value has to be either 0 / 1 or
FALSE / TRUE (0 or FALSE if item is not bought). Columns should be
named.}

\item{m_sup}{User specified minimum support}

\item{m_conf}{User specified minimum confidence}
}
\value{
Returns an object of class \code{associationrules}
}
\description{
Mines association rules using the Apriori Algorithm.
The function demands a transactionmatrix, user specified minimum
support and minimum confidence and returns generated association rules.
First it transforms the data using \code{create_transactionmatrix}, then
it uses \code{freq_items} to find frequent itemsets and finally
generates association rules using \code{rules}. These functions can also
be accessed seperately.
The function \code{apriorimining} returns an object of class
associationrules, which can be analysed using \code{show} and \code{summary}.
}
