% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_associationrules.R
\docType{class}
\name{AssociationRules-class}
\alias{AssociationRules-class}
\alias{show,AssociationRules-method}
\alias{summary,AssociationRules-method}
\title{AssociationRules}
\usage{
\S4method{show}{AssociationRules}(object)

\S4method{summary}{AssociationRules}(object)
}
\arguments{
\item{object}{Object of class \code{AssociationRules}}
}
\description{
The S4 class \code{AssociationRules} characterises the generated
association rules using five different slots. An object of class
\code{AssociationRules} can be analysed using \code{show} and \code{summary}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Shows the number of the generated association rules

\item \code{summary}: Returns a data.frame of all generated
association rules and their respective quality measurements
}}

\section{Slots}{

\describe{
\item{\code{items}}{Character vector containing the item names of the transaction matrix}

\item{\code{antecedent}}{Each antecedent of every generated association rule
in the shape of a sparse matrix (ngCMatrix)}

\item{\code{consequent}}{Each consequent of every generated association rule
in the shape of a sparse matrix (ngCMatrix)}

\item{\code{measurements}}{Quality measurements (support, confidence, lift)
for every generated association rule}

\item{\code{m_values}}{User specified minimum values for support and confidence}
}}

