% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_items.R
\name{freq_items}
\alias{freq_items}
\title{Find frequent itemsets}
\usage{
freq_items(input, m_sup)
}
\arguments{
\item{input}{Binary matrix containing transaction data, with rows
representing transactions and columns representing items. Can be
either logical or numeric, every value has to be either 0 / 1 or
FALSE / TRUE (0 or FALSE if item is not bought). Columns should be
named. Also takes an object class \code{transactiondata}.}

\item{m_sup}{User specified minimum support}
}
\value{
Returns an object of class \code{frequentsets}
}
\description{
\code{freq_item} finds every frequent itemset depending
on the chosen minimum support. The function demands a transactionsmatrix
and a user specified minimum support.
}
