% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_frequentsets.R
\docType{class}
\name{frequentsets-class}
\alias{frequentsets-class}
\alias{show,frequentsets-method}
\alias{summary,frequentsets-method}
\alias{plot,frequentsets-method}
\title{Frequentsets}
\usage{
\S4method{show}{frequentsets}(object)

\S4method{summary}{frequentsets}(object)

\S4method{plot}{frequentsets}(x)
}
\arguments{
\item{object}{object of class frequentsets}

\item{x}{object of class \code{frequentsets}}
}
\description{
The S4 class \code{frequentsets} characterizes the frequently
occuring itemsets using four different slots. Objects of the class
\code{frequentsets} can be analyzed with \code{show},
\code{summary} and \code{plot}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Shows the total number of the frequently occuring
itemsets and how often the different sets sizes occur.

\item \code{summary}: Provides an overview of the frequently occuring itemsets

\item \code{plot}: Plots a histogram of the support distribution for each set size
}}

\section{Slots}{

\describe{
\item{\code{sets}}{Every generated itemset summarized in the shape of an sparsematrix
(ngCMatrix)}

\item{\code{supports}}{Numeric vector containing the calculated support for the
respective sets}

\item{\code{items}}{Character vector containing the itemnames of the transactionmatrix}

\item{\code{minsup}}{User specified minimum support}
}}

