% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groceries.R
\docType{data}
\name{Groceries}
\alias{Groceries}
\title{Data: Groceries}
\format{Object of class \code{transactiondata}}
\source{
obtained from the \code{arules} package, created by Michael Hahsler,
Kurt Hornik and Thomas Reutterer.
}
\usage{
Groceries
}
\description{
The data consists of 9835 transactions and 169 items. In the present package,
data is in the shape of a sparse matrix
}
\references{
Michael Hahsler, Kurt Hornik, and Thomas Reutterer (2006) Implications
of probabilistic data modeling for mining association rules. In M. Spiliopoulou,
R. Kruse, C. Borgelt, A. Nuernberger, and W. Gaul, editors, From Data and Information
Analysis to Knowledge Engineering, Studies in Classification, Data Analysis, and
Knowledge Organization, pages 598–605. Springer-Verlag.
}
\keyword{datasets}
