% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_transactiondata.R
\name{itemtail}
\alias{itemtail}
\title{Test minimum support}
\usage{
itemtail(object, support, absolute = FALSE, rare = FALSE)
}
\arguments{
\item{object}{Object of class TransactionData}

\item{support}{user specified minimum support}

\item{absolute}{logical}

\item{rare}{logical}
}
\description{
Test minimum support
}
