% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{rules}
\alias{rules}
\title{Generate association rules}
\usage{
rules(itemsets, m_conf)
}
\arguments{
\item{itemsets}{Object of class frequentsets}

\item{m_conf}{User specified minimum confidence}
}
\value{
Returs an object of class \code{associationrules}
}
\description{
The function \code{rules} generates all possible association rules.
The function demands the frequently occuring itemsets and a user specified
minimum confidence.
}
