% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_transactiondata.R
\docType{class}
\name{TransactionData-class}
\alias{TransactionData-class}
\alias{show,TransactionData-method}
\alias{summary,TransactionData-method}
\alias{itemtail,TransactionData-method}
\alias{plot,TransactionData-method}
\alias{tryminimum,TransactionData-method}
\title{TransactionData}
\usage{
\S4method{show}{TransactionData}(object)

\S4method{summary}{TransactionData}(object)

\S4method{itemtail}{TransactionData}(object, support, absolute = FALSE,
  rare = FALSE)

\S4method{plot}{TransactionData}(x)

\S4method{tryminimum}{TransactionData}(object)
}
\arguments{
\item{object}{Object of class \code{TransactionData}}

\item{support}{User specified minimum support}

\item{absolute}{Logical}

\item{rare}{Logical}

\item{x}{Object of class \code{TransactionData}}
}
\description{
The S4 class \code{TransactionData} characterises the
entered transaction matrix using two different slots. An object of the
class \code{TransactionData} can be analysed with \code{show}, \code{summary}
,\code{plot}, \code{itemtail} and \code{tryminimum}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Shows number of transactions and items
in the transaction matrix

\item \code{summary}: Returns data.frame of each item in the
transaction matrix and their frequencies

\item \code{itemtail}: Returns data.frame of either the most freqeunt
single items or the most rare single items.
User can choose wether the occurence of the items is displayed in terms
of absolute or relative values.

\item \code{plot}: Bar plot of the most frequently occuring
items displaying their absolute frequencies

\item \code{tryminimum}: Provides the number of frequent items for four
different minimum support values, giving an impression of the influence of
the minimum support
}}

\section{Slots}{

\describe{
\item{\code{data}}{Transaction matrix in the shape of a sparsematrix (ngCMatrix)}

\item{\code{items}}{Character vector containing the item names of the
trancastion matrix}
}}

