% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_transactiondata.R
\name{tryminimum}
\alias{tryminimum}
\title{Test influence of different minimum supports}
\usage{
tryminimum(object)
}
\arguments{
\item{object}{Object of class TransActiondata}
}
\description{
Test influence of different minimum supports
}
