% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apriorimining.R
\name{apriorimining}
\alias{apriorimining}
\title{Determine association rules}
\usage{
apriorimining(input, m_sup, m_conf)
}
\arguments{
\item{input}{Binary matrix containing transaction data, with rows
representing transactions and columns representing items. Can be
either logical or numeric, every value has to be either 0 / 1 or
FALSE / TRUE (0 or FALSE if item is not bought). Columns should be
named.}

\item{m_sup}{User specified minimum support}

\item{m_conf}{User specified minimum confidence}
}
\value{
Returns an object of class \code{AssociationRules}
}
\description{
Mines association rules using the Apriori Algorithm.
The function demands a transaction matrix, user specified minimum
support and minimum confidence and returns generated association rules.
First \code{apriorimining} transforms the data into an
object of class \code{TransactionData} using \code{create_transactionmatrix},
then the function passes this object
to \code{freq_items} in order to find frequent item sets. Finally \code{apriorimining}
generates association rules from those frequent item sets using \code{rules}.
These functions can also be accessed seperately.
The function \code{apriorimining} returns an object of class
AssociationRules, which can be analysed using \code{show} and \code{summary}.
}
\examples{
\dontrun{
## Load data
data(Groceries)
## Mine rules
x <- apriorimining(Groceries, m_sup = 0.05, m_conf = 0.3)
## use methods
summary(x)
}
}
