% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_associationrules.R
\docType{class}
\name{associationrules-class}
\alias{associationrules-class}
\alias{show,associationrules-method}
\alias{summary,associationrules-method}
\title{Associationrules}
\usage{
\S4method{show}{associationrules}(object)

\S4method{summary}{associationrules}(object)
}
\arguments{
\item{object}{object of class \code{associationrules}}
}
\description{
The S4 class \code{associationrules} characterizes the generated
association rules using four different slots. An object of class
\code{associationrules} can be analysed using \code{show} and \code{summary}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Shows the number of the generated association rules

\item \code{summary}: Returns a data.frame of all generated
association rules and their respective
quality measurements
}}

\section{Slots}{

\describe{
\item{\code{items}}{Character vector containing the itemnames of the transactionmatrix}

\item{\code{antecedent}}{Each antecedent of every generated association rule
in the shape of an sparsematrix (ngCMatrix)}

\item{\code{consequent}}{Each consequent of every generated association rule
in the shape of an sparsematrix (ngCMatrix)}

\item{\code{measurements}}{Quality measurements (support, confidence, lift)
for every generated association rule}
}}

