% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_transactionmatrix.R
\name{create_transaction}
\alias{create_transaction}
\title{Transform the transactionmatrix}
\usage{
create_transaction(input)
}
\arguments{
\item{input}{Binary matrix containing transaction data, with rows
representing transactions and columns representing items. Can be
either logical or numeric, every value has to be either 0 / 1 or
FALSE / TRUE (0 or FALSE if item is not bought). Columns should be
named.}
}
\value{
Returns an object of class \code{transactiondata}
}
\description{
Transforms a given transactionmatrix into an object
of class transactiondata, enabling the user to examine the transactiondata
with provided methodes.
}
