% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_items.R
\name{freq_items}
\alias{freq_items}
\title{Find frequent item sets}
\usage{
freq_items(input, m_sup)
}
\arguments{
\item{input}{Binary matrix containing transaction data, with rows
representing transactions and columns representing items. Can be
either logical or numeric, every value has to be either 0 / 1 or
FALSE / TRUE (0 or FALSE if item is not bought). Columns should be
named. \code{freq_item} also takes an object class \code{TransactionData}.}

\item{m_sup}{User specified minimum support}
}
\value{
Returns an object of class \code{FrequentSets}
}
\description{
\code{freq_item} finds every frequent item set depending
on the chosen minimum support. The function demands a transaction matrix
and a user specified minimum support.
}
