% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_frequentsets.R
\docType{class}
\name{FrequentSets-class}
\alias{FrequentSets-class}
\alias{show,FrequentSets-method}
\alias{summary,FrequentSets-method}
\alias{plot,FrequentSets-method}
\title{FrequentSets}
\usage{
\S4method{show}{FrequentSets}(object)

\S4method{summary}{FrequentSets}(object)

\S4method{plot}{FrequentSets}(x)
}
\arguments{
\item{object}{Object of class FrequentSets}

\item{x}{Object of class \code{FrequentSets}}
}
\description{
The S4 class \code{FrequentSets} characterises the frequently
occuring item sets using four different slots. Objects of the class
\code{FrequentSets} can be analysed with \code{show},
\code{summary} and \code{plot}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Shows the total number of the frequently occuring
item sets and how often the different sets sizes occur.

\item \code{summary}: Provides an overview of the frequently occuring item sets

\item \code{plot}: Plots a bar plot of the frequencies of the different
set sizes.
}}

\section{Slots}{

\describe{
\item{\code{sets}}{Every generated item set summarised in the shape of a sparsematrix
(ngCMatrix)}

\item{\code{supports}}{Numeric vector containing the calculated support for the
respective sets}

\item{\code{items}}{Character vector containing the item names of the transaction matrix}

\item{\code{minsup}}{User specified minimum support}
}}

