% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{rules}
\alias{rules}
\title{Generate association rules}
\usage{
rules(itemsets, m_conf)
}
\arguments{
\item{itemsets}{Object of class FrequentSets}

\item{m_conf}{User specified minimum confidence}
}
\value{
Returs an object of class \code{AssociationRules}
}
\description{
The function \code{rules} generates all possible association rules.
The function demands the frequently occuring item sets and a user specified
minimum confidence.
}
\examples{
\dontrun{
## create FrequentSets object
x <- freq_items(data, m_sup = 0.3)
## find rules
y <- rules(x, m_conf = 0.2)
## inspect
summary(y)
}
}
