% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_transactiondata.R
\docType{class}
\name{transactiondata-class}
\alias{transactiondata-class}
\alias{show,transactiondata-method}
\alias{summary,transactiondata-method}
\alias{itemtail,transactiondata-method}
\alias{plot,transactiondata-method}
\alias{tryminimum,transactiondata-method}
\title{Transactiondata}
\usage{
\S4method{show}{transactiondata}(object)

\S4method{summary}{transactiondata}(object)

\S4method{itemtail}{transactiondata}(object, support, absolute = FALSE,
  rare = FALSE)

\S4method{plot}{transactiondata}(x)

\S4method{tryminimum}{transactiondata}(object)
}
\arguments{
\item{object}{object of class \code{transactiondata}}

\item{support}{user specified minimum support}

\item{absolute}{logical}

\item{rare}{logical}

\item{x}{object of class \code{transactiondata}}
}
\description{
The S4 class \code{transactiondata} characterizes the
entered transactionmatrix using two different slots. An object of the
class \code{transactiondata} can be analyzed with \code{show}, \code{summary}
,\code{plot}, \code{itemtail} and \code{tryminimum}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Shows number of transactions and items in the transactionmatrix

\item \code{summary}: Returns data.frame of each item in the transactionmatrix and
their frequencies

\item \code{itemtail}: returns data.frame of either the most freqeunt
single items or the most rare single items.
User can choose wether the occurence of the items is displayed in terms
of absolute or relative values.

\item \code{plot}: barplot of the most frequently occuring
items displaying their absolute frequencies

\item \code{tryminimum}: provides the number of frequent items for four
different minimum support values, giving an impression of the influence of
the minimum support
}}

\section{Slots}{

\describe{
\item{\code{data}}{Transaction data in the shape of an sparsematrix (ngCMatrix)}

\item{\code{items}}{Character vector containing the itemnames of the
trancastionmatrix}
}}

